Move PartType enum, fix naming, cleanup
