+++ PartsInventory/Models/Enums.cs
@@ -141,5 +141,58 @@ namespace PartsInventory.Models.Enums
       Keylock = 1111,
       Navigation = 1112,
    };
+
+   public enum PartType
+   {
+      Unknown = 0,
+
+      Resistor,
+      Capacitor,
+      Inductor,
+      IC,
+      Arduino,
+      Battery,
+      CapacitorNetwork,
+      ResistorNetwork,
+      Diode,
+      DiodeNetwork,
+      Display,
+      Fuse,
+      FerriteBead,
+      Fiducial,
+      Filter,
+      InfraredDiode,
+      Connector,
+      Jumper,
+      Relay,
+      Speaker,
+      Motor,
+      Microphone,
+      OptoIsolator,
+      PowerSupply,
+      Transistor,
+      Thermistor,
+      Varistor,
+      Switch,
+      Transformer,
+      Thermocouple,
+      TestPoint,
+      Tuner,
+      VacuumTube,
+      VoltageRegulator,
+      Potentiometer,
+      Crystal,
+      Oscillator,
+      BridgeRectifier,
+      Attenuator,
+      DelayLine,
+      Hardware,
+      DirectionalCoupler,
+      Socket,
+
+      USBConnector,
+      SenseResistor,
+      RotaryEncoder,
+   }
    #endregion
 }
